Refactor solvers into modular package
